KAGE-GPT2/Src/requirements.txt
absl-py
alabaster==0.7.12
allennlp==0.9.0
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
#aws-xray-sdk==0.95
#awscli==1.16.132
Babel==2.6.0
#boto==2.49.0
#boto3==1.17.3
#botocore==1.12.122
cachetools
certifi
cffi==1.12.2
chardet==3.0.4
checksumdir==1.2.0
Click
colorama==0.3.9
colorlog==4.0.2
colorlover==0.3.0
conllu==1.3.1
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
Cython==0.29.21
cytoolz==0.9.0.1
dataclasses==0.6
deap==1.2.2
decorator==4.4.0
dill==0.2.9
docker==3.7.1
docker-pycreds==0.4.0
docutils==0.14
easydict==1.9
ecdsa==0.13
editdistance==0.5.3
embeddings==0.0.8
filelock
flaky==3.5.3
Flask==1.0.2
Flask-Cors==3.0.7
flatbuffers==1.10
ftfy==5.5.1
funcsigs==1.0.2
future
fuzzywuzzy==0.18.0
gast==0.2.2
gevent==1.3.6
google-auth
google-auth-oauthlib
google-pasta==0.2.0
googledrivedownloader==0.4
greenlet==0.4.15
grpcio
gym==0.12.1
h5py==2.9.0
idna
imagesize==1.1.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.10.1
jmespath==0.9.4
joblib
jsondiff==1.1.1
#jsonnet
jsonpatch==1.26
jsonpickle==1.1
jsonpointer==2.0
jsonschema==3.0.1
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
kiwisolver
Markdown
MarkupSafe==1.1.1
matplotlib
mock==2.0.0
more-itertools==6.0.0
moto==1.3.4
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbformat==4.4.0
nltk==3.4
numpy
numpydoc==0.8.0
oauthlib
opencv-python==4.1.2.30
opt-einsum==3.3.0
overrides==1.9
packaging
pandas==0.24.2
parsimonious==0.8.0
pbr==5.1.3
Pillow
#pkg-resources==0.0.0
plac==0.9.6
plotly==3.7.1
pluggy==0.9.0
preshed==2.0.1
protobuf
psutil==5.7.2
py==1.8.0
pyaml==18.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.8.0
pydash==4.7.4
pyglet==1.3.2
Pygments==2.3.1
pyparsing
pyrsistent==0.14.11
pytest==4.3.1
python-dateutil
python-jose==2.0.2
python-Levenshtein==0.12.0
pytorch-ignite==0.2.0
pytorch-pretrained-bert==0.6.1
pytorch-transformers==1.1.0
pytz==2019.1
PyYAML
pyzmq==19.0.2
quadprog==0.1.7
ray==0.6.5
redis==3.2.1
regex
requests
requests-oauthlib==1.3.0
responses==0.10.6
retrying==1.3.3
rsa
s3transfer==0.3.0
sacremoses
scikit-learn
scipy==1.2.1
sentencepiece
simplejson==3.16.0
singledispatch==3.4.0.3
six
snowballstemmer==1.2.1
spacy
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
#tensorboard
#tensorboard-plugin-wit
#tensorboardX==2.1
#tensorflow==1.15.4
#tensorflow-estimator==1.15.1
termcolor==1.1.0
thinc
tokenizers
toolz==0.9.0
#torch==1.6.0
#torchfile==0.1.0
tornado==6.0.4
tqdm
traitlets==4.3.2
typing==3.6.6
ujson==1.35
Unidecode==1.0.23
urllib3
visdom==0.1.8.9
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug
word2number==1.1
wrapt
xmltodict==0.12.0
xvfbwrapper==0.2.9
#absl-py==0.14.1
#cachetools==4.2.4
#certifi==2021.5.30
charset-normalizer==2.0.6
#click==8.0.1
cycler==0.10.0
dataclasses==0.6
easydict==1.9
#filelock==3.3.0
#future==0.18.2
#google-auth==1.35.0
#google-auth-oauthlib==0.4.6
#grpcio==1.41.0
#idna==3.2
importlib-metadata==4.8.1
#joblib==1.0.1
jsonnet==0.17.0
#kiwisolver==1.3.2
#Markdown==3.3.4
#matplotlib==3.4.3
#numpy==1.21.2
#oauthlib==3.1.1
#packaging==21.0
#Pillow==8.3.2
#protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
#pyparsing==2.4.7
#python-dateutil==2.8.2
#regex==2021.9.30
#requests==2.26.0
requests-oauthlib==1.3.0
#rsa==4.7.2
#sacremoses==0.0.46
#scikit-learn==1.0
scipy==1.2.1
#sentencepiece==0.1.91
#six==1.16.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.0.0
#tokenizers==0.9.3
#torch==1.7.0
#tqdm==4.62.3
transformers==3.5.1
typing-extensions==3.10.0.2
#urllib3==1.26.7
#Werkzeug==2.0.1
zipp==3.6.0